cleanup docs, address code review comments
